UPD:merging_calls
